+++ READMD.md
@@ -1,0 +1,8 @@
+# auto-canary
+一个基于华为云CCE集群接口开发的内部生产部署平台，主要用于帮助简化灰度发布上线流程中的各种运维逻辑，如版本上线前对网关流量的控制，简化生产流程。
+## feature
+- 支持内部用户管理
+- 支持灰度版本更新到线上版本的流量切换逻辑
+- 支持灰度版本更新的流量切换逻辑
+
+
