--- doc/.copyright
@@ -1,1 +1,0 @@
-[加入编程导航](https://www.code-nav.cn/) 入门捷径+交流答疑+项目实战+求职指导，帮你自学编程不走弯路
