+++ pom.xml
@@ -1,8 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <!-- @author <a href=#">一只小小丑</a> -->
-    <!-- @from 中科院软件所 -->
     <modelVersion>4.0.0</modelVersion>
     <parent>
         <groupId>org.springframework.boot</groupId>
@@ -10,8 +8,8 @@
         <version>2.6.4</version>
         <relativePath/> <!-- lookup parent from repository -->
     </parent>
-    <groupId>com.yupi</groupId>
-    <artifactId>user-center-backend</artifactId>
+    <groupId>com.iscas</groupId>
+    <artifactId>auto-canary</artifactId>
     <version>0.0.1-SNAPSHOT</version>
     <name>user-center-backend</name>
     <description>user-center-backend</description>
@@ -71,6 +69,12 @@
             <artifactId>junit</artifactId>
             <version>4.13.2</version>
             <scope>test</scope>
+        </dependency>
+
+        <dependency>
+            <groupId>com.huaweicloud.sdk</groupId>
+            <artifactId>huaweicloud-sdk-cce</artifactId>
+            <version>3.1.78</version>
         </dependency>
     </dependencies>
