--- sql/.author
@@ -1,1 +1,0 @@
-作_者 [程序员_鱼皮](https://github.com/liyupi)
