+++ sql/create_table.sql
@@ -3,10 +3,10 @@
 # @from 中科院软件所
 
 -- 创建库
-create database if not exists yupi;
+create database if not exists auto_canary;
 
 -- 切换库
-use yupi;
+use auto_canary;
 
 # 用户表
 create table user
@@ -30,5 +30,5 @@ create table user
     comment '用户';
 
 # 导入示例用户
-INSERT INTO yupi.user (username, userAccount, avatarUrl, gender, userPassword, phone, email, userStatus, createTime, updateTime, isDelete, userRole, planetCode) VALUES ('鱼皮', 'yupi', 'https://himg.bdimg.com/sys/portraitn/item/public.1.e137c1ac.yS1WqOXfSWEasOYJ2-0pvQ', null, 'b0dd3697a192885d7c055db46155b26a', null, null, 0, '2023-08-06 14:14:22', '2023-08-06 14:39:37', 0, 1, '1');
+INSERT INTO auto_canary.user (username, userAccount, avatarUrl, gender, userPassword, phone, email, userStatus, createTime, updateTime, isDelete, userRole, planetCode) VALUES ('user1', 'auto_canary', 'https://himg.bdimg.com/sys/portraitn/item/public.1.e137c1ac.yS1WqOXfSWEasOYJ2-0pvQ', null, 'b0dd3697a192885d7c055db46155b26a', null, null, 0, '2023-08-06 14:14:22', '2023-08-06 14:39:37', 0, 1, '1');
