+++ src/main/java/com/iscas/autoCanary/contant/UserConstant.java
@@ -1,0 +1,29 @@
+package com.iscas.autoCanary.contant;
+
+/**
+ * 用户常量
+ *
+ * @author <a href=#">一只小小丑</a>
+ * @from 中科院软件所
+ */
+public interface UserConstant {
+
+    /**
+     * 用户登录态键
+     */
+    String USER_LOGIN_STATE = "userLoginState";
+
+    //  ------- 权限 --------
+
+    /**
+     * 默认权限
+     */
+    int DEFAULT_ROLE = 0;
+
+
+    /**
+     * 管理员权限
+     */
+    int ADMIN_ROLE = 1;
+
+}
