+++ src/main/java/com/iscas/autoCanary/controller/CCEController.java
@@ -1,0 +1,66 @@
+package com.iscas.autoCanary.controller;
+
+import com.iscas.autoCanary.common.BaseResponse;
+import com.iscas.autoCanary.common.ErrorCode;
+import com.iscas.autoCanary.common.ResultUtils;
+import com.iscas.autoCanary.contant.UserConstant;
+import com.iscas.autoCanary.exception.BusinessException;
+import com.iscas.autoCanary.model.domain.User;
+import com.iscas.autoCanary.model.domain.request.UserRegisterRequest;
+import org.apache.commons.lang3.StringUtils;
+import org.springframework.web.bind.annotation.*;
+
+import javax.servlet.http.HttpServletRequest;
+
+/**
+ * @author windpo
+ * @description 实现对集群自动化操作的逻辑
+ */
+@RestController
+@RequestMapping("/cce")
+public class CCEController {
+    @PostMapping("/updateOld/cutFlow")
+    public BaseResponse cutOldFlow(HttpServletRequest request) {
+        Object userObj = request.getSession().getAttribute(UserConstant.USER_LOGIN_STATE);
+        User currentUser = (User) userObj;
+        if (currentUser == null) {
+            throw new BusinessException(ErrorCode.NOT_LOGIN);
+        }
+        // todo 修改configMap，切断稳定版本流量
+        return ResultUtils.success(null);
+    }
+
+    @PostMapping("/updateOld/resumeFlow")
+    public BaseResponse resumeOldFlow(HttpServletRequest request) {
+        Object userObj = request.getSession().getAttribute(UserConstant.USER_LOGIN_STATE);
+        User currentUser = (User) userObj;
+        if (currentUser == null) {
+            throw new BusinessException(ErrorCode.NOT_LOGIN);
+        }
+        // todo 修改configMap，恢复稳定版本流量
+        return ResultUtils.success(null);
+    }
+
+    @PostMapping("/updateNew/cutFlow")
+    public BaseResponse cutNewFlow(HttpServletRequest request) {
+        Object userObj = request.getSession().getAttribute(UserConstant.USER_LOGIN_STATE);
+        User currentUser = (User) userObj;
+        if (currentUser == null) {
+            throw new BusinessException(ErrorCode.NOT_LOGIN);
+        }
+        // todo 修改configMap，切断灰度版本流量
+        return ResultUtils.success(null);
+    }
+
+
+    @PostMapping("/updateNew/resumeFlow")
+    public BaseResponse resumeNewFlow(HttpServletRequest request) {
+        Object userObj = request.getSession().getAttribute(UserConstant.USER_LOGIN_STATE);
+        User currentUser = (User) userObj;
+        if (currentUser == null) {
+            throw new BusinessException(ErrorCode.NOT_LOGIN);
+        }
+        // todo 修改configMap，恢复灰度版本流量
+        return ResultUtils.success(null);
+    }
+}
