+++ src/main/java/com/iscas/autoCanary/exception/BusinessException.java
@@ -1,0 +1,50 @@
+package com.iscas.autoCanary.exception;
+
+
+import com.iscas.autoCanary.common.ErrorCode;
+
+/**
+ * 自定义异常类
+ *
+ * @author <a href=#">一只小小丑</a>
+ * @from 中科院软件所
+ */
+public class BusinessException extends RuntimeException {
+
+    /**
+     * 异常码
+     */
+    private final int code;
+
+    /**
+     * 描述
+     */
+    private final String description;
+
+    public BusinessException(String message, int code, String description) {
+        super(message);
+        this.code = code;
+        this.description = description;
+    }
+
+    public BusinessException(ErrorCode errorCode) {
+        super(errorCode.getMessage());
+        this.code = errorCode.getCode();
+        this.description = errorCode.getDescription();
+    }
+
+    public BusinessException(ErrorCode errorCode, String description) {
+        super(errorCode.getMessage());
+        this.code = errorCode.getCode();
+        this.description = description;
+    }
+
+    public int getCode() {
+        return code;
+    }
+
+
+    public String getDescription() {
+        return description;
+    }
+}
