+++ src/main/java/com/iscas/autoCanary/exception/GlobalExceptionHandler.java
@@ -1,0 +1,31 @@
+package com.iscas.autoCanary.exception;
+
+import com.iscas.autoCanary.common.BaseResponse;
+import com.iscas.autoCanary.common.ErrorCode;
+import com.iscas.autoCanary.common.ResultUtils;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+/**
+ * 全局异常处理器
+ *
+ * @author <a href=#">一只小小丑</a>
+ * @from 中科院软件所
+ */
+@RestControllerAdvice
+@Slf4j
+public class GlobalExceptionHandler {
+
+    @ExceptionHandler(BusinessException.class)
+    public BaseResponse<?> businessExceptionHandler(BusinessException e) {
+        log.error("businessException: " + e.getMessage(), e);
+        return ResultUtils.error(e.getCode(), e.getMessage(), e.getDescription());
+    }
+
+    @ExceptionHandler(RuntimeException.class)
+    public BaseResponse<?> runtimeExceptionHandler(RuntimeException e) {
+        log.error("runtimeException", e);
+        return ResultUtils.error(ErrorCode.SYSTEM_ERROR, e.getMessage(), "");
+    }
+}
