+++ src/main/java/com/iscas/autoCanary/mapper/UserMapper.java
@@ -1,0 +1,16 @@
+package com.iscas.autoCanary.mapper;
+
+import com.iscas.autoCanary.model.domain.User;
+import com.baomidou.mybatisplus.core.mapper.BaseMapper;
+
+/**
+ * 用户 Mapper
+ *
+ * @author <a href=#">一只小小丑</a>
+ * @from 中科院软件所
+ */
+public interface UserMapper extends BaseMapper<User> {
+
+}
+
+
