+++ src/main/java/com/iscas/autoCanary/model/domain/request/UserRegisterRequest.java
@@ -1,0 +1,37 @@
+package com.iscas.autoCanary.model.domain.request;
+
+import lombok.Data;
+
+import java.io.Serializable;
+
+/**
+ * 用户注册请求体
+ *
+ * @author <a href=#">一只小小丑</a>
+ * @from 中科院软件所
+ */
+@Data
+public class UserRegisterRequest implements Serializable {
+
+    private static final long serialVersionUID = 3191241716373120793L;
+
+    /**
+     * 用户账号
+     */
+    private String userAccount;
+
+    /**
+     * 用户密码
+     */
+    private String userPassword;
+
+    /**
+     * 校验密码
+     */
+    private String checkPassword;
+
+    /**
+     * 星球编号
+     */
+    private String planetCode;
+}
