--- src/main/java/com/yupi/usercenter/UserCenterApplication.java
@@ -1,24 +1,0 @@
-package com.yupi.usercenter;
-
-// [加入编程导航](https://www.code-nav.cn/) 深耕编程提升【两年半】、国内净值【最高】的编程社群、用心服务【20000+】求学者、帮你自学编程【不走弯路】
-
-import org.mybatis.spring.annotation.MapperScan;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-/**
- * 启动类
- *
- * @author 一只小小丑
- * @from
- */
-@SpringBootApplication
-@MapperScan("com.yupi.usercenter.mapper")
-public class UserCenterApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(UserCenterApplication.class, args);
-    }
-
-}
-
