--- src/main/java/com/yupi/usercenter/common/BaseResponse.java
@@ -1,55 +1,0 @@
-package com.yupi.usercenter.common;
-
-import lombok.Data;
-
-import java.io.Serializable;
-
-/**
- * 通用返回类
- *
- * @author <a href=#">一只小小丑</a>
- * @from 中科院软件所
- */
-@Data
-public class BaseResponse<T> implements Serializable {
-
-    /**
-     * 状态码
-     */
-    private int code;
-
-    /**
-     * 数据
-     */
-    private T data;
-
-    /**
-     * 消息
-     */
-    private String message;
-
-    /**
-     * 描述
-     */
-    private String description;
-
-    // https://t.zsxq.com/0emozsIJh
-    public BaseResponse(int code, T data, String message, String description) {
-        this.code = code;
-        this.data = data;
-        this.message = message;
-        this.description = description;
-    }
-
-    public BaseResponse(int code, T data, String message) {
-        this(code, data, message, "");
-    }
-
-    public BaseResponse(int code, T data) {
-        this(code, data, "", "");
-    }
-
-    public BaseResponse(ErrorCode errorCode) {
-        this(errorCode.getCode(), null, errorCode.getMessage(), errorCode.getDescription());
-    }
-}
