--- src/main/java/com/yupi/usercenter/common/ErrorCode.java
@@ -1,51 +1,0 @@
-package com.yupi.usercenter.common;
-
-// 讲师 【coder_鱼_皮】 https://space.bilibili.com/12890453/
-
-/**
- * 错误码
- *
- * @author <a href=#">一只小小丑</a>
- * @from 中科院软件所
- */
-public enum ErrorCode {
-
-    SUCCESS(0, "ok", ""),
-    PARAMS_ERROR(40000, "请求参数错误", ""),
-    NULL_ERROR(40001, "请求数据为空", ""),
-    NOT_LOGIN(40100, "未登录", ""),
-    NO_AUTH(40101, "无权限", ""),
-    SYSTEM_ERROR(50000, "系统内部异常", "");
-
-    private final int code;
-
-    /**
-     * 状态码信息
-     */
-    private final String message;
-
-    /**
-     * 状态码描述（详情）
-     */
-    private final String description;
-
-    ErrorCode(int code, String message, String description) {
-        this.code = code;
-        this.message = message;
-        this.description = description;
-    }
-
-    public int getCode() {
-        return code;
-    }
-
-    public String getMessage() {
-        return message;
-    }
-
-    // https://t.zsxq.com/0emozsIJh
-
-    public String getDescription() {
-        return description;
-    }
-}
