--- src/main/java/com/yupi/usercenter/contant/UserConstant.java
@@ -1,30 +1,0 @@
-package com.yupi.usercenter.contant;
-
-/**
- * 用户常量
- *
- * @author <a href=#">一只小小丑</a>
- * @from 中科院软件所
- */
-public interface UserConstant {
-
-    /**
-     * 用户登录态键
-     */
-    String USER_LOGIN_STATE = "userLoginState";
-
-    //  ------- 权限 --------
-
-    /**
-     * 默认权限
-     */
-    int DEFAULT_ROLE = 0;
-
-    // https://space.bilibili.com/12890453/
-
-    /**
-     * 管理员权限
-     */
-    int ADMIN_ROLE = 1;
-
-}
