--- src/main/java/com/yupi/usercenter/exception/GlobalExceptionHandler.java
@@ -1,33 +1,0 @@
-package com.yupi.usercenter.exception;
-
-import com.yupi.usercenter.common.BaseResponse;
-import com.yupi.usercenter.common.ErrorCode;
-import com.yupi.usercenter.common.ResultUtils;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-
-/**
- * 全局异常处理器
- *
- * @author <a href=#">一只小小丑</a>
- * @from 中科院软件所
- */
-@RestControllerAdvice
-@Slf4j
-public class GlobalExceptionHandler {
-
-    @ExceptionHandler(BusinessException.class)
-    public BaseResponse<?> businessExceptionHandler(BusinessException e) {
-        log.error("businessException: " + e.getMessage(), e);
-        return ResultUtils.error(e.getCode(), e.getMessage(), e.getDescription());
-    }
-
-    //https://github.com/liyupi
-
-    @ExceptionHandler(RuntimeException.class)
-    public BaseResponse<?> runtimeExceptionHandler(RuntimeException e) {
-        log.error("runtimeException", e);
-        return ResultUtils.error(ErrorCode.SYSTEM_ERROR, e.getMessage(), "");
-    }
-}
