--- src/main/java/com/yupi/usercenter/mapper/UserMapper.java
@@ -1,18 +1,0 @@
-package com.yupi.usercenter.mapper;
-
-import com.yupi.usercenter.model.domain.User;
-import com.baomidou.mybatisplus.core.mapper.BaseMapper;
-
-/**
- * 用户 Mapper
- *
- * @author <a href=#">一只小小丑</a>
- * @from 中科院软件所
- */
-public interface UserMapper extends BaseMapper<User> {
-
-}
-
-
-// [编程知识星球](https://yupi.icu) 零基础快速入门编程，不走弯路！30+ 原创学习路线和专栏、1000+ 编程精华文章、500+ 编程学习指南、20T+ 编程资源汇总
-
