--- src/main/java/com/yupi/usercenter/model/domain/request/UserLoginRequest.java
@@ -1,31 +1,0 @@
-package com.yupi.usercenter.model.domain.request;
-
-// [编程学习交流圈](https://www.code-nav.cn/) 快速入门编程不走弯路！30+ 原创学习路线和专栏、500+ 编程学习指南、1000+ 编程精华文章、20T+ 编程资源汇总
-
-import lombok.Data;
-
-import java.io.Serializable;
-
-/**
- * 用户登录请求体
- *
- * @author <a href=#">一只小小丑</a>
- * @from 中科院软件所
- */
-@Data
-public class UserLoginRequest implements Serializable {
-
-    private static final long serialVersionUID = 3191241716373120793L;
-
-    /**
-     * 用户账号
-     */
-    private String userAccount;
-
-    /**
-     * 用户密码
-     */
-    private String userPassword;
-
-    // https://www.code-nav.cn/
-}
