--- src/main/java/com/yupi/usercenter/model/domain/request/UserRegisterRequest.java
@@ -1,39 +1,0 @@
-package com.yupi.usercenter.model.domain.request;
-
-import lombok.Data;
-
-import java.io.Serializable;
-
-/**
- * 用户注册请求体
- *
- * @author <a href=#">一只小小丑</a>
- * @from 中科院软件所
- */
-@Data
-public class UserRegisterRequest implements Serializable {
-
-    private static final long serialVersionUID = 3191241716373120793L;
-
-    /**
-     * 用户账号
-     */
-    private String userAccount;
-
-    /**
-     * 用户密码
-     */
-    private String userPassword;
-
-    /**
-     * 校验密码
-     */
-    private String checkPassword;
-
-    /**
-     * 星球编号
-     */
-    private String planetCode;
-}
-
-// [加入编程导航](https://yupi.icu) 入门捷径+交流答疑+项目实战+求职指导，帮你自学编程不走弯路
