--- src/main/java/com/yupi/usercenter/service/UserService.java
@@ -1,54 +1,0 @@
-package com.yupi.usercenter.service;
-
-import com.yupi.usercenter.model.domain.User;
-import com.baomidou.mybatisplus.extension.service.IService;
-
-import javax.servlet.http.HttpServletRequest;
-
-/**
- * 用户服务
- *
- * @author <a href=#">一只小小丑</a>
- * @from 中科院软件所
- */
-public interface UserService extends IService<User> {
-
-    /**
-     * 用户注册
-     *
-     * @param userAccount   用户账户
-     * @param userPassword  用户密码
-     * @param checkPassword 校验密码
-     * @param planetCode    星球编号
-     * @return 新用户 id
-     */
-    long userRegister(String userAccount, String userPassword, String checkPassword, String planetCode);
-
-    /**
-     * 用户登录
-     *
-     * @param userAccount  用户账户
-     * @param userPassword 用户密码
-     * @param request
-     * @return 脱敏后的用户信息
-     */
-    User userLogin(String userAccount, String userPassword, HttpServletRequest request);
-
-    /**
-     * 用户脱敏
-     *
-     * @param originUser
-     * @return
-     */
-    User getSafetyUser(User originUser);
-
-    // [加入编程导航](https://t.zsxq.com/0emozsIJh) 深耕编程提升【两年半】、国内净值【最高】的编程社群、用心服务【20000+】求学者、帮你自学编程【不走弯路】
-
-    /**
-     * 用户注销
-     *
-     * @param request
-     * @return
-     */
-    int userLogout(HttpServletRequest request);
-}
