--- src/main/resources/.yupi
@@ -1,1 +1,0 @@
-开发者 [coder_yupi](https://yupi.icu/)
