+++ src/main/resources/application-prod.yml
@@ -1,17 +1,13 @@
-# 线上配置文件
-# @author <a href=#">一只小小丑</a>
-# @from 中科院软件所
+#线上配置文件
 spring:
   # DataSource Config
   datasource:
     driver-class-name: com.mysql.jdbc.Driver
     url: 'jdbc:mysql://sh-cynosdbmysql-grp-98pxrcoq.sql.tencentcdb.com:29164/user_center?useSSL=false'
-    username: self
-    password: liyupi66!
+    username: {your_remote_root} #ex:root
+    password: {your_remote_password} #ex:123456
   # session 失效时间
   session:
     timeout: 86400
 server:
   address: 0.0.0.0
-
-# [加入学习圈](https://t.zsxq.com/0emozsIJh) 从 0 到 1 项目实战，经验拉满！10+ 原创项目手把手教程、1000+ 项目经验笔记、7 日项目提升训练营、60+ 编程经验分享直播
