+++ src/main/resources/application.yaml
@@ -1,15 +1,12 @@
-# 公共配置文件
-# @author <a href=#">一只小小丑</a>
-# @from 中科院软件所
 spring:
   application:
-    name: user-center-backend
+    name: auto-canary
   # DataSource Config
   datasource:
     driver-class-name: com.mysql.jdbc.Driver
-    url: jdbc:mysql://localhost:3306/yupi
-    username: root
-    password: liyifan81349219
+    url: jdbc:mysql://localhost:3306/auto_canary
+    username: {your_user}   #example:root
+    password: {your_password}   #example:123456
   # session 失效时间
   session:
     timeout: 86400
@@ -25,4 +22,6 @@ mybatis-plus:
       logic-delete-field: isDelete # 全局逻辑删除的实体字段名(since 3.3.0,配置后可以忽略不配置步骤2)
       logic-delete-value: 1 # 逻辑已删除值(默认为 1)
       logic-not-delete-value: 0 # 逻辑未删除值(默认为 0)
+HUAWEICLOUD_SDK_AK: xxxxxxxxxxx
+HUAWEICLOUD_SDK_SK: xxxxxxxxxxx
