+++ src/main/resources/mapper/UserMapper.xml
@@ -1,12 +1,10 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<!-- @author <a href=#">一只小小丑</a> -->
-<!-- @from 中科院软件所 -->
 <!DOCTYPE mapper
         PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
         "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.yupi.usercenter.mapper.UserMapper">
+<mapper namespace="com.iscas.autoCanary.mapper.UserMapper">
 
-    <resultMap id="BaseResultMap" type="com.yupi.usercenter.model.domain.User">
+    <resultMap id="BaseResultMap" type="com.iscas.autoCanary.model.domain.User">
         <id property="id" column="id" jdbcType="BIGINT"/>
         <result property="username" column="username" jdbcType="VARCHAR"/>
         <result property="userAccount" column="userAccount" jdbcType="VARCHAR"/>
@@ -31,5 +29,4 @@
         userRole,planetCode
     </sql>
 
-<!--    https://yupi.icu/-->
 </mapper>
