--- src/test/java/com/yupi/usercenter/UserCenterApplicationTests.java
@@ -1,37 +1,0 @@
-package com.yupi.usercenter;
-
-
-import org.junit.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.util.DigestUtils;
-
-import javax.annotation.Resource;
-import java.nio.charset.StandardCharsets;
-import java.security.MessageDigest;
-import java.security.NoSuchAlgorithmException;
-
-/**
- * 启动类测试
- *
- * @author <a href=#">一只小小丑</a>
- * @from 中科院软件所
- */
-@SpringBootTest
-class UserCenterApplicationTests {
-
-
-    @org.junit.jupiter.api.Test
-    void testDigest() throws NoSuchAlgorithmException {
-        String newPassword = DigestUtils.md5DigestAsHex(("abcd" + "mypassword").getBytes());
-        System.out.println(newPassword);
-    }
-
-
-    @Test
-    void contextLoads() {
-
-    }
-
-}
-
-// [编程知识星球](https://yupi.icu) 连接万名编程爱好者，一起优秀！20000+ 小伙伴交流分享、100+ 各方向编程交流群、40+ 大厂嘉宾一对一答疑、4000+ 编程问答参考
